Created script for saving variant data with control data easily locally and on Sherlock.
